chore: move test from lib to tests
